refactor common logic out of astyanax specific classes
