+++ blueflood-core/src/main/java/com/rackspacecloud/blueflood/io/CassandraModel.java
@@ -3,8 +3,8 @@ package com.rackspacecloud.blueflood.io;
 import com.netflix.astyanax.model.ColumnFamily;
 import com.netflix.astyanax.serializers.LongSerializer;
 import com.netflix.astyanax.serializers.StringSerializer;
-import com.rackspacecloud.blueflood.io.serializers.LocatorSerializer;
-import com.rackspacecloud.blueflood.io.serializers.SlotStateSerializer;
+import com.rackspacecloud.blueflood.io.serializers.astyanax.LocatorSerializer;
+import com.rackspacecloud.blueflood.io.serializers.astyanax.SlotStateSerializer;
 import com.rackspacecloud.blueflood.rollup.Granularity;
 import com.rackspacecloud.blueflood.service.Configuration;
 import com.rackspacecloud.blueflood.service.CoreConfig;
