+++ blueflood-core/src/main/java/com/rackspacecloud/blueflood/io/serializers/astyanax/LocatorSerializer.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package com.rackspacecloud.blueflood.io.serializers;
+package com.rackspacecloud.blueflood.io.serializers.astyanax;
 
 import com.rackspacecloud.blueflood.types.Locator;
 import com.google.common.base.Charsets;
