+++ blueflood-core/src/main/java/com/rackspacecloud/blueflood/io/serializers/json/GaugeRollupDeserializer.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package com.rackspacecloud.blueflood.io.serializers;
+package com.rackspacecloud.blueflood.io.serializers.json;
 
 import com.rackspacecloud.blueflood.types.BluefloodGaugeRollup;
 import org.codehaus.jackson.JsonParser;
