+++ blueflood-core/src/main/java/com/rackspacecloud/blueflood/io/serializers/metrics/package-info.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2016 Rackspace
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+/**
+ * This package is to hold all of the driver agnostic serializer/deserializer
+ * classes. It is important to isolate the driver specific things into
+ * wrapper classes such as those created for Astyanax and Datastax.
+ *
+ * The classes here mainly deal with the wire format of the objects to
+ * be serialized. The objects to be serialized are mainly Metrics related
+ * objects, such as Set, Gauge, Histogram, Timer, Counter, Enum, etc.
+ */
+package com.rackspacecloud.blueflood.io.serializers.metrics;
