+++ blueflood-core/src/test/java/com/rackspacecloud/blueflood/io/serializers/astyanax/StringMetadataSerializerTest.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package com.rackspacecloud.blueflood.io.serializers;
+package com.rackspacecloud.blueflood.io.serializers.astyanax;
 
 import org.junit.Test;
